fix: round 2 review — double-frees, memory leaks, and type fixes in mongo wrapper
- Fix BsonValue move double-free: custom move ctor/assign deep-copies
  bson_value_t and zeroes source, preventing double-free on destroy
- Fix bson_t* leak in all 5 BsonDocument factory methods:
  bson_new_* returns heap-allocated struct; add bson_destroy + bson_free
- Fix null-reply crash in 3 CommandSimple variants: reply→RawBson()
  without null check
- Fix MongoStream factory double-free: base stream ownership transferred
  to C API; call ReleaseStream() to prevent wrapper from double-freeing
- Fix CreateBulkOperation silently ignoring ordered/session_raw params
- Fix l_bson_iter_gc missing pointer nullification after delete
- Fix Tell() missing const qualifier on MongoGridFsFile
- Fix GetMaxStalenessSeconds truncating int64_t to int
- Fix SetChunkSize taking signed int32_t instead of uint32_t
- Fix incomplete-type risk: out-of-line MongoServerDescription dtor
  since unique_ptr<MongoHostList> needs complete type

+++ src/runtime/database/mongo/mongo_settings.cc
@@ -80,9 +80,9 @@ bool MongoReadPrefs::AddTag(const BsonDocument& tag) {
     return true;
 }
 
-int MongoReadPrefs::GetMaxStalenessSeconds() const {
+int64_t MongoReadPrefs::GetMaxStalenessSeconds() const {
     return impl_ && impl_->prefs
-        ? static_cast<int>(mongoc_read_prefs_get_max_staleness_seconds(impl_->prefs)) : 0;
+        ? mongoc_read_prefs_get_max_staleness_seconds(impl_->prefs) : 0;
 }
 
 void MongoReadPrefs::SetMaxStalenessSeconds(int max_staleness_seconds) {
